requirements.txt
attrs==23.1.0
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1648854242877/work
cattrs==23.1.2
certifi==2022.9.24
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1656782892388/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
conda==22.9.0
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1663666418037/work
coremltools==7.0
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1657174295336/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
distlib==0.3.6
exceptiongroup==1.1.3
filelock==3.8.0
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1661381467342/work
fsspec==2023.10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
Jinja2==3.1.2
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1657953165042/work
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1637843342785/work/libmambapy
mamba @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1637843342785/work/mamba
MarkupSafe==2.1.3
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1639359043885/work
mpmath==1.3.0
munkres==1.1.4
networkx==3.2.1
nodeenv==1.8.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1662888942478/work
opencv-python==4.6.0.66
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1660385974755/work
platformdirs==2.5.2
protobuf==3.20.3
pyaml==23.9.7
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1649384941891/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1663846997386/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
pyright==1.1.333
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
PyYAML==6.0.1
regex==2023.6.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1661872987712/work
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1653464548430/work
scikit-spatial @ file:///home/conda/feedstock_root/build_artifacts/scikit-spatial_1633614169160/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sympy==1.12
tiktoken==0.4.0
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
torch==2.1.0
torchvision==0.16.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1656938068018/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
typing_extensions==4.8.0
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1649111974586/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
virtualenv==20.16.4
